PLATFORM-3575 changing formulary request field include_plan to include_plans
ENTERPRISE-281: completed s3 deploy script
