+++ .gitignore
@@ -13,7 +13,6 @@ test/version_tmp
 tmp
 *.DS_STORE
 build/
-BACKUP/
 .cache
 
 # YARD artifacts
