+++ build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+cd /app
+bundle install
+middleman build
