+++ requirements.txt
@@ -1,0 +1,5 @@
+argparse==1.2.1
+boto3==1.3.1
+botocore==1.4.19
+requests==2.2.1
+virtualenv==1.11.2
